feat: implemented basic typescript
