+++ package.json
@@ -42,5 +42,16 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/react-redux": "^7.1.11",
+    "@typescript-eslint/eslint-plugin": "^4.6.1",
+    "@typescript-eslint/parser": "^4.6.1",
+    "eslint": "^7.13.0",
+    "eslint-config-htmlacademy": "^0.6.0",
+    "eslint-plugin-react": "^7.21.5",
+    "redux-devtools-extension": "^2.13.8",
+    "ts-jest": "^26.4.4",
+    "ts-loader": "^8.0.11"
   }
 }
