+++ src/App.test.js
@@ -1,7 +1,8 @@
-import { render, screen } from '@testing-library/react';
+import * as React from "react";
+import {render, screen} from '@testing-library/react';
 import App from './App';
 
-test('renders learn react link', () => {
+test(`renders learn react link`, () => {
   render(<App />);
   const linkElement = screen.getByText(/learn react/i);
   expect(linkElement).toBeInTheDocument();
