+++ src/App.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import Users from './components/UsersComponent';
 import './App.css';
