+++ src/redux/actions/users-action.ts
@@ -1,0 +1,8 @@
+import * as type from '../types';
+
+export function getUsers(users: any) {
+  return {
+    type: type.GET_USERS.REQUESTED,
+    payload: users
+  };
+}
