--- src/redux/actions/users.js
@@ -1,8 +1,0 @@
-import * as type from '../types';
-
-export function getUsers(users) {
-    return {
-        type: type.GET_USERS.REQUESTED,
-        payload: users
-    }
-}
