--- src/redux/reducers/index.js
@@ -1,8 +1,0 @@
-import {combineReducers} from "redux";
-import users from './users';
-
-const rootReducers = combineReducers({
-    users: users,
-})
-
-export default rootReducers;
