+++ src/redux/reducers/root-reducer.ts
@@ -1,0 +1,8 @@
+import {combineReducers} from "redux";
+import users from './users';
+
+const rootReducer = combineReducers({
+  users,
+});
+
+export default rootReducer;
