--- src/redux/sagas/index.js
@@ -1,8 +1,0 @@
-import { all } from 'redux-saga/effects'
-import userSaga from './userSaga'
-
-export default function* rootSaga() {
-  yield all([
-    userSaga(),
-  ])
-}
