+++ src/redux/sagas/root-saga.ts
@@ -1,0 +1,8 @@
+import {all} from 'redux-saga/effects';
+import userSaga from './user-saga';
+
+export default function* rootSaga() {
+  yield all([
+    userSaga(),
+  ]);
+}
