+++ src/redux/store.ts
@@ -1,0 +1,16 @@
+import {applyMiddleware, createStore} from 'redux';
+import createSagaMiddleware from 'redux-saga';
+import rootReducer from './reducers/root-reducer';
+import rootSaga from './sagas/root-saga';
+import {composeWithDevTools} from 'redux-devtools-extension';
+
+const sagaMiddleware = createSagaMiddleware();
+
+const store = createStore(
+    rootReducer,
+    composeWithDevTools(applyMiddleware(sagaMiddleware),
+    ));
+
+sagaMiddleware.run(rootSaga);
+
+export default store;
