--- src/redux/types.js
@@ -1,5 +1,0 @@
-export const GET_USERS = {
-    REQUESTED: 'GET_USERS_REQUESTED',
-    SUCCESS: 'GET_USERS_SUCCESS',
-    FAILED: 'GET_USERS_FAILED'
-}
