+++ src/redux/types.ts
@@ -1,0 +1,11 @@
+interface GetUsers {
+  REQUESTED: string;
+  SUCCESS: string;
+  FAILED: string;
+}
+
+export const GET_USERS: GetUsers = {
+  REQUESTED: `GET_USERS_REQUESTED`,
+  SUCCESS: `GET_USERS_SUCCESS`,
+  FAILED: `GET_USERS_FAILED`
+};
